Use canonical internal URIs with trailing slashes
Hugo uses 'directory'-like URIs for pages which end with a slash. Lack
of a slash is fine, but it generates and unnecessary HTTP redirect with
the  web-server configuration we use for the docserver. This can lead
to worse SEO ratings which we don't like.

Search and replace URIs to add slashes where necessary. I know, this is
stupid and silly for humans to do this, but our Google overlords demand
this offering.