Use CC BY-NC-ND 4.0 for content and Apache 2.0 for code (#3)
Since the documentation is mostly prose, open-source licenses for
software do not really apply to it well.

In particular, GNU Affero GPL is a copyleft license, and it can be
interpreted in a way that treats modified code snippets as derivative
work. This will require our users to distribute their software under
the same terms, if they use a widely populare software development
methodology "copy-paste examples from the docs".

Our open-source products are usually licensed under Apache 2.0,
so let's apply the same license to all code snippets by default.

As for the prose, let's use a license specifically designed for
content: Creative Commons in its BY-NC-ND flavor. That allows hosting
and citing our work with proper attribution, prevents silly games with
charging for access (we display public docs for free), and also prevents
derivative works by default (please contact us if you want to improve
the documentation in some way). Along with that, it's well-known license
which should be enforceable if necessary.

There is no precedent for CC usage. The current docs do not specify
any open license, so the default copyright law applies (assigning it
to Cossack Labs Limited). We can go with that as well if we change the
wording a bit.