Use staging-friendly docserver URLs
Replace (almost) all direct links to https://docs.cossacklabs.com
with domain-relative paths. That way the links will work with the
staging instance of the documentation server correctly, forwarding
the users to the staging instance.

Currently, at one place we have a link that is a verbatim part of
a warning message. That one is left intact (themis-server.md).