ObjCThemis feature guide: overhaul
Update the structure of the guide similar to ThemisPP and others. Most
the changes involve Secure Cell interface which has gone through some
changes. Also, now we don't need to differentiate between CocoaPods
and Carthage: since Themis 0.13 they have the same import syntax.

Another interesting thing is that ObjCThemis has callback API for
Secure Session, but it is not working. So it is commented out.