SwiftThemis feature guide: overhaul
For the most part, Swift guide is a translation of the Objective-C one.
However, there are some tweaks since Swift does error handling in a
different way. We prefer "guard" and "try" to explicit "do" blocks
so that the code is more terse. The users will figure out what to do
with Swift exceptions.