Provide links to API documentation
Where possible, link to actual automatically generated API docs.
Currently, those are Go, Rust, and suddenly Ruby.

Other languages do not have generated and hosted API docs yet. Include
commented out links to fictional /docs/themis/api/ directory on the
documentation server as a placeholder. Some day it might be live.

Since we currently do not host versioned documentation, link to the
'latest' version on the API docs. This has a benefit of not having
to keep those links constantly updated. However, if the documentation
is versioned then it will make more sense to link to API docs for
a matching version.