Use staging-friendly docserver URLs (#6)
* Use staging-friendly docserver URLs

Replace (almost) all direct links to https://docs.cossacklabs.com
with domain-relative paths. That way the links will work with the
staging instance of the documentation server correctly, forwarding
the users to the staging instance.

Currently, at one place we have a link that is a verbatim part of
a warning message. That one is left intact (themis-server.md).

* Host image with NSIS instructions in this repo

The image at https://docs.cossacklabs.com/files/wiki/cossack-labs-themis-win-installer.png
is currently hosted somewhere on the old doc server. Mirror it here as
well so that the link does not break, and the image is still visible
in local builds.

* Avoid verbatim docserver link

After all, drop the verbatim link from Themis Server docs as well.
It might actually be outdated, but we don't have time to verify.