Feature guides: Objective-C, Swift, PHP, Python, Ruby, Rust (#8)
* ObjCThemis feature guide: overhaul

Update the structure of the guide similar to ThemisPP and others. Most
the changes involve Secure Cell interface which has gone through some
changes. Also, now we don't need to differentiate between CocoaPods
and Carthage: since Themis 0.13 they have the same import syntax.

Another interesting thing is that ObjCThemis has callback API for
Secure Session, but it is not working. So it is commented out.

* SwiftThemis feature guide: overhaul

For the most part, Swift guide is a translation of the Objective-C one.
However, there are some tweaks since Swift does error handling in a
different way. We prefer "guard" and "try" to explicit "do" blocks
so that the code is more terse. The users will figure out what to do
with Swift exceptions.

* PHPThemis feature guide: overhaul

Well, this is short one. PHPThemis is not supported well and has not
changed much. Just replace API references with actual code samples.

Secure Session interface is particularly unwieldy, but it is what it is.

Secure Comparator is not supported by PHPThemis, we leave a stub.

* PyThemis feature guide: overhaul

Update PyThemis guide to match the common template as well.

PyThemis is different in that its Secure Message has a one-off API for
the sign-verify mode. Also, Secure Session callback API is wonky.

Also, since we're supporting only Python 3 now, drop all mentions of
Python 2 from the text.

* JsThemis, WasmThemis guides: drop Secure Session modes

JavaScript supports only buffer-aware API, drop descriptions of modes
that have dead links. Secure Session usage is described in a single
section.

* RbThemis feature guide: overhaul

Update Ruby guide as well. Like Python, Secure Message here has a bit
different API. Other parts of the API also have a bit different naming.

* RustThemis feature guide: overhaul

Update Rust guide too. It's mostly similar to C++. The only quirk is
that Secure Session in Rust supports only EC keys at the moment.

Also, drop contribution instructions. There is a section for that
elsewhere in the docs.

* Note missing callback-oriented API of Secure Session

Like in RbThemis, add a note explaining why callback API is not
available in some languages. In some cases we have decided that
it's not a good fit and in some cases it's simply not ready yet.

* Favor error handling in Context Imprint mode

Examples for other modes show error handling in decryption. Since
Context Imprint mode is not able to report errors, they must be checked
manually by the user. Make an emphasis on that, and avoid wording
"looks correct" since it's a bit imprecise.

* Use the same wording for Secure Comparator results

The same structure:

    if (matching secrets) {
        // shared secrets match
    }

and the same wording in comments.

* Provide links to API documentation

Where possible, link to actual automatically generated API docs.
Currently, those are Go, Rust, and suddenly Ruby.

Other languages do not have generated and hosted API docs yet. Include
commented out links to fictional /docs/themis/api/ directory on the
documentation server as a placeholder. Some day it might be live.

Since we currently do not host versioned documentation, link to the
'latest' version on the API docs. This has a benefit of not having
to keep those links constantly updated. However, if the documentation
is versioned then it will make more sense to link to API docs for
a matching version.