Acra key management (#17)
* Basic structure of "Key management"

Begin by salvaging information from current documentation server.
However, now we are not constrained by the previous limitations
and can introduce more sections.

* Key management cover page

This key management guide is a compilation of existing docs, some early
drafts of documentation for key store v2, and pages from staging version
about key rotation.

* Describe Acra key inventory
* Describe troubleshooting
* Describe key store versions
* Describe key generation
* Describe key destruction
* Describe key rotation
* Describe key store versions
* Describe key management concisely for index

* Describe making backups

Well, the story with backups is complicated. Tools are different for
various versions. We also might consider telling that it's sufficient
to back up the ".acrakeys" directory, but -- duh! -- that should be
obvious if you are going to make encrypted backups yourself.

* Describe key store migration

Co-authored-by: vixentael <vixentael@users.noreply.github.com>