Consistently use "keystore" as a single word
Like in the code, replace all uses with a single word for the sake of
consistency. Spelling "keystore" as a single word more closely resembles
command-line syntax, is similar to how we use "keypair", and avoids
weird connotations with stores as in "marketplaces". If it were called
"key storage" maybe it would be okay to spell it separately...