Basic skeleton structure of Themis specs
Start describing Themis cryptograhy and data structure design in
a series of technical specifications. Hopefully, this will make
reimplementing Themis in different platforms easier as well as
simplifies security audit and analysis. It may also be beneficial
to new developers to serve as a map for the codebase (or rather,
as a textbook which you apply to the codebase).

This commit adds the cover page and defines the basic structure
of the specification chapter. It will take a while to write so
unfinished sections get a notice instead of content. The pages
are already available so that the hyperlinks work right away.