Asymmetric keys: EC
Describe the structure of EC keys used by Themis. This is the preferred
key kind, useful for its compactness. There are some minor details in
the format used by Themis which sometimes catch developers off guard.
Hopefully, detailed description of the format with dissected examples
helps to avoid such misunderstandings in the future.