Themis specification: keys and keypairs (#25)
* Basic skeleton structure of Themis specs

Start describing Themis cryptograhy and data structure design in
a series of technical specifications. Hopefully, this will make
reimplementing Themis in different platforms easier as well as
simplifies security audit and analysis. It may also be beneficial
to new developers to serve as a map for the codebase (or rather,
as a textbook which you apply to the codebase).

This commit adds the cover page and defines the basic structure
of the specification chapter. It will take a while to write so
unfinished sections get a notice instead of content. The pages
are already available so that the hyperlinks work right away.

* Common data: Soter container

Soter container is pervasive in Themis. It's used to store asymmetric
keys as well as Secure Session protocol messages. Describe the layout
and specific algorithm choices. Provide a code illustration so that
it's more accessible to software developers.

* Symmetric keys and passphrases

Secure Cell relies on symmetric keys and passphrases for security. While
Themis does not have a strict format for keys, we still should describe
this fact and give some useful advice on generating them. The same goes
for the passphrases.

* Asymmetric keys: index page

Themis uses a couple of asymmetric key kinds which have similar
structure and at times are interchangeable. However, they are still
distinct so details are delegated to their own pages.

* Asymmetric keys: EC

Describe the structure of EC keys used by Themis. This is the preferred
key kind, useful for its compactness. There are some minor details in
the format used by Themis which sometimes catch developers off guard.
Hopefully, detailed description of the format with dissected examples
helps to avoid such misunderstandings in the future.

* Asymmetric keys: RSA

Similar to EC keys, describe RSA keys -- an alternate cryptosystem
supported by Secure Message. While RSA keys are not compact -- as you
can readily see in the examples sectio -- they have some performance
benefits over EC keys, and may be a better choice in some cases.

* Drop unnecessary "under construction" banners

Remove the banners from the pages with content, leaving them on the
cryptosystem pages which are yet to be written.

* Be explicit about code snippet language

Make sure to say that the code snippets are in Go. We use Go since
it is one of the primary languages in Cossack Labs. Go also has nice
support of cryptography in its standard library, and is more or less
readable to anyone who can read C.

* Rename "Crypto{graphy => graphic} library"

This wording seems to be way more popular across the documentation.

* Explain preference of EC keys over RSA

Add a short blurb on the asymmetric key page which explains why Themis
supports several asymmetric key types, but don't go into much detail.
This will be explored on the Secure Message page.

* Improve wording of passphrase description

Explain that passphrases have less entropy per byte compared to keys
instead of calling them "weird symmetric keys".

* Rename "Crypto{graphic => graphy} library" back

After all, "cryptography library" seems to be more popular on the
Internet so let's go with that wording.

* Apply suggestions from code review

Co-authored-by: vixentael <vixentael@users.noreply.github.com>