Secure Cell layout (#31)
* Secure Cell layout

Describe data layout of Secure Cell in detail. Explain how the
authentication tokens are structured, for both symmetric key and
passphrase cases, and describe the Soter algorithm descriptor too
which is an integral part of the autentication token (and used by
some other cryptosystems as well).

* Secure Cell example data

Provide dissected example data to illustrate the layout described above.
It is possible to cross-check the data using the interactive simulator
or command-line utilities.

* Clarify interpretation of test inputs

They are intended to be passed as is on the terminal to the command-line
utilities. Mention relevant utilities (they have the same naming pattern
in all languages). Also, reorder the lines to follow the order in which
the utilities take their arguments.

* Mention NIST SP rationale for parameter lengths

The choice of IV, auth tag, and salt lengths is technically arbitrary
but it influences the performance and security of the overall scheme.
NIST Special Publications analyze algorithms and provide recommendations
which we adhere to.