Convert "spec/common" into a directory
There will be more and probably more sections added to this page and
I'd like to keep it managable.

For one, I'd like the anchors to have some meaningul names, instead of
"#example-1" referring to Soter algorithm descriptor example which will
change if you dare to move the section.

At the same time, it's tiresome to ensure that each section has unique
name that fits in one line. Sure, you can name it "Example: Soter
algorithm descriptor" or "Soter algorithm example", but this removes
a very useful structure: each page having an "Example" section.

So... Since with the new doc server we're moving away from the "monopage
concept", let's do this to the common section of the spec.