Use reference implementation instead of example
Now that we have a reference implementation in our repository, let's
link to it for the implementation details and replace the snippet with
something shorter (which still invokes the same code). The reader should
be able to copy the snippet, run it with "go run", and see the output
which corresponds to whatever is there in the spec prose.