Split "spec/common" sections into their own pages (#34)
* Convert "spec/common" into a directory

There will be more and probably more sections added to this page and
I'd like to keep it managable.

For one, I'd like the anchors to have some meaningul names, instead of
"#example-1" referring to Soter algorithm descriptor example which will
change if you dare to move the section.

At the same time, it's tiresome to ensure that each section has unique
name that fits in one line. Sure, you can name it "Example: Soter
algorithm descriptor" or "Soter algorithm example", but this removes
a very useful structure: each page having an "Example" section.

So... Since with the new doc server we're moving away from the "monopage
concept", let's do this to the common section of the spec.

* Split common sections into their own pages

Just like that, move each top-level section onto its own page, and the
index page now serves as an index, or a table of contents.

* Update internal URLs

Instead of using anchors, use the new page names.

I could have added backup compatibility anchors to the "common" index
page, but since it's so small the users wouldn't have an issue locating
the next link they should follow.

It would be cool to do such redirects automatically, but we still don't
have the machinery for anchors.

* Update index page title

Right now this section of the spec describes only 'data structures'
but later on Soter KDF description will be added here, maybe with
some friends. Let's rename the page for accuracy.

Note that the "sidebar name" stays the same so that it fits in one line.