Reference implementation of Soter container (#35)
* Reference implementation of Soter container

Okay, here's my stupid idea. In addition to specs with prose, provide
users with a reference implementation of whatever is described. Some
people are better at understanding concepts through code. Plus, this
provides an "executable specification" and tests. I have found this
invaluable when writing Secure Cell spec: that caught quite a few
mistakes and misconceptions that I had about its implementation.

(History-savvy people here could have found parallels between this and
the concept of "literate programming". Indeed, if only Knuth's ideas
caught up, this would have been a perfect opportunity for that approach.
Unfortunately, it's not really popular to have good tooling.)

The reference implementation is written in Go. Aside from an obvious
reasons having to do with Cossack Labs being a Go shop, this choice has
a number of other benefits:

  - very good standard library, with a lot of cryptographic primitives
    available which will come very handy later

  - readable to most programmers who are familiar with C syntax

  - almost immediately executable if you copy-paste the sample code,
    thanks to "push to GitHub to publish a package" approach

  - reference implementation can actually reuse its own code for
    different cryptosystems

Unfortunately, you can't run this on Go Playground without much hassle,
but that would have been amazing.

Now, note that "reference implementation" is different from the
"official GoThemis source code". The overriding objective here is to
illustrate key points in Themis algorithms. Readability and correctness
are the main focus of the reference implementation. Performance,
usability, good error reporting, feature completeness -- all of this can
be sacrificed for the sake of brevity and clarity of the code.

Themis Core's C code is anything but easy to understand cryptography.
Since even its maintainers make mistakes when deciphering code paths
leading to arcane OpenSSL invocations, any cryptography audit would be
pretty expensive for it. Hopefully, this reference implementation can
serve as a better way to understand the cryptography behind Themis.

Now... Here's an implementation of Soter container for starters.

* Use reference implementation instead of example

Now that we have a reference implementation in our repository, let's
link to it for the implementation details and replace the snippet with
something shorter (which still invokes the same code). The reader should
be able to copy the snippet, run it with "go run", and see the output
which corresponds to whatever is there in the spec prose.