Touch up WasmThemis installation instructions (#225)
* themis/wasm: Provide Node.js stub example for webpack 5

The example we have was written in the times of webpack 3. It's still
actual for webpack 4, but webpack 5 has invented a new syntax.

Also, it seems the latest version of WasmThemis needs more of those
polyfills. Mention them.

* themis/wasm: Use Emscripten 3.0.0 in build instructions

And provide a link to the repository where we keep the current version.
Older versions on WasmThemis might need older versions, but if you're
trying to build an older version, you should be competent to figure out
which version you need.